+++ src/main/java/server/quest/actions/MesoAction.java
@@ -40,7 +40,6 @@ public class MesoAction extends AbstractQuestAction {
         processData(data);
     }
 
-
     @Override
     public void processData(Data data) {
         mesos = DataTool.getInt(data);
@@ -56,10 +55,10 @@ public class MesoAction extends AbstractQuestAction {
             chr.gainMeso(gain, true, false, true);
         } else {
             if (!YamlConfig.config.server.USE_QUEST_RATE) {
-                chr.gainMeso(gain * chr.getMesoRate(), true, false, true);
+                chr.gainMeso((int) (gain * chr.getMesoRate()), true, false, true);
             } else {
                 chr.gainMeso(gain * chr.getQuestMesoRate(), true, false, true);
             }
         }
     }
-} 
+}
